# docs/openapi_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Content Automation ERP API",
    "description": "Production-grade API-First Control Plane for YouTube Automation & Durable Workflows",
    "version": "1.0.0"
  },
  "paths": {
...
    "/videos/{idea_id}/status": {
      "get": {
        "tags": [
          "Videos"
        ],
        "summary": "Get Video Status",
        "operationId": "get_video_status_videos__idea_id__status_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{idea_id}/assemble": {
      "post": {
        "tags": [
          "Videos"
        ],
        "summary": "Assemble Video Shots",
        "operationId": "assemble_video_shots_videos__idea_id__assemble_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssembleVideoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssembleVideoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{idea_id}": {
      "get": {
        "tags": [
          "Videos"
        ],
        "summary": "Get Video Record",
        "operationId": "get_video_record_videos__idea_id__get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoRecordSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{idea_id}/render-payload": {
      "get": {
        "tags": [
          "Videos"
        ],
        "summary": "Get Render Payload",
        "operationId": "get_render_payload_videos__idea_id__render_payload_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{idea_id}/generate": {
      "post": {
        "tags": [
          "Videos"
        ],
        "summary": "Generate Video",
        "description": "Triggers 10-second video generation for an idea.\nIf dry_run=True, validates prompt and builds payload without launching browser.",
        "operationId": "generate_video_videos__idea_id__generate_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Dry Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/workflows": {
      "post": {
        "tags": [
          "Workflow Builder"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "Workflow Builder"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}/versions/draft": {
      "post": {
        "tags": [
          "Workflow Builder"
        ],
        "summary": "Create Draft Version",
        "operationId": "create_draft_version_workflows__workflow_id__versions_draft_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftVersionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/versions/{version_id}/steps": {
      "put": {
        "tags": [
          "Workflow Builder"
        ],
        "summary": "Update Draft Steps",
        "operationId": "update_draft_steps_workflows_versions__version_id__steps_put",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftStepsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/versions/{version_id}": {
      "get": {
        "tags": [
          "Workflow Builder"
        ],
        "summary": "Get Workflow Version",
        "operationId": "get_workflow_version_workflows_versions__version_id__get",
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssembleVideoRequest": {
        "properties": {
          "shot_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shot Paths"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          }
        },
        "type": "object",
        "required": [
          "shot_paths"
        ],
        "title": "AssembleVideoRequest"
      },
      "AssembleVideoResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "total_shots": {
            "type": "integer",
            "title": "Total Shots"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status",
          "total_shots"
        ],
        "title": "AssembleVideoResponse"
      },
...
      "CreateDraftVersionRequest": {
        "properties": {
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          },
          "changelog": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changelog"
          }
        },
        "type": "object",
        "title": "CreateDraftVersionRequest"
      },
...
      "CreateWorkflowRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Unique workflow name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "shorts"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateWorkflowRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDraftStepsRequest": {
        "properties": {
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "UpdateDraftStepsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoRecordSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "idea_id",
          "status"
        ],
        "title": "VideoRecordSchema"
      },
...
    }
  }
}